7th commit
New class, UserServices created to handle the login success requirement and to act as a database by utilizing the initializeAccount method.
-added
App.java - username and password textbox output removed.
New class created Account.java which handles the class account and creates the attributes and constructor in creating an account for the login app - this also contained the Getters which includes the getPassword ethod which grabs a copy of the password to prevent exposure.